Fixing circular import error caused by setting up the test model for API development.
